requirements.txt
python-dotenv==1.0.0
requests==2.25.1
beautifulsoup4==4.9.3
selenium==4.15.2
undetected-chromedriver==3.5.3
pycryptodome==3.10.1
inky[rpi,example-depends]==1.5.0
Pillow==10.1.0
